refactor script to make it easier to read and add in another loop, to do both All Recipes and Meat-Free
